JIRA #JSS-3321: improvements when setting repository root paths
Keeping in sync the information set in the project preferences (root
paths) and the single project resource properties.
Added a new attribute in the dedicated extension point: we need an
utility class that helps in performing additional operations whenever a
context switch happens.
Improved the existing APIs working with the repository root paths and
the interaction with the JRIOManager instances.